copilotkit-pydantic/first-party-mcp-servers/microsoft365/requirements.txt
# Microsoft 365 MCP Server Requirements

# MCP Framework
fastmcp>=0.3.0

# Microsoft Graph SDK
msgraph-sdk>=1.0.0

# Authentication (Azure Identity)
azure-identity>=1.15.0
azure-core>=1.30.0

# HTTP Client
requests>=2.31.0

# Caching
cachetools>=5.3.0

# Web Server
uvicorn[standard]>=0.27.0

# Type Safety & Validation
pydantic>=2.0.0

# Date/Time handling
python-dateutil>=2.8.2

# Optional: Local file processing
openpyxl>=3.1.0        # Excel file processing
python-docx>=1.1.0     # Word document processing

